Test refactoring: Extract hbcluster provisioning and checking for first values
