+++ flink-connectors/flink-connector-hbase/src/test/java/org/apache/flink/connector/hbase/source/HBaseSourceITCase.java
@@ -6,16 +6,13 @@ import org.apache.flink.api.common.serialization.AbstractDeserializationSchema;
 import org.apache.flink.api.java.tuple.Tuple2;
 import org.apache.flink.connector.hbase.source.hbasemocking.DemoIngester;
 import org.apache.flink.connector.hbase.source.hbasemocking.DemoSchema;
-import org.apache.flink.connector.hbase.source.hbasemocking.TestClusterStarter;
+import org.apache.flink.connector.hbase.source.hbasemocking.HBaseTestClusterUtil;
 import org.apache.flink.core.execution.JobClient;
 import org.apache.flink.streaming.api.datastream.DataStream;
 import org.apache.flink.streaming.api.environment.StreamExecutionEnvironment;
 import org.apache.flink.util.Collector;
 
 import org.apache.hadoop.hbase.client.Put;
-import org.junit.After;
-import org.junit.AfterClass;
-import org.junit.BeforeClass;
 import org.junit.Test;
 
 import java.io.IOException;
@@ -26,32 +23,7 @@ import java.util.concurrent.TimeUnit;
 import static org.junit.Assert.assertArrayEquals;
 
 /** Tests the most basic use cases of the source with a mocked HBase system. */
-public class HBaseSourceITCase {
-
-    /** For debug purposes. Allows to run the test quickly without starting a fresh cluster */
-    public static final boolean USE_EXISTING_CLUSTER = false;
-
-    /** Shadowed from org.apache.flink.test.util.SuccessException. */
-    public static class SuccessException extends RuntimeException {}
-
-    @BeforeClass
-    public static void setup() {
-        if (!USE_EXISTING_CLUSTER) {
-            TestClusterStarter.startCluster();
-        }
-        assert TestClusterStarter.isClusterAlreadyRunning();
-    }
-
-    @AfterClass
-    public static void teardown() throws IOException {
-        if (!USE_EXISTING_CLUSTER) TestClusterStarter.shutdownCluster();
-    }
-
-    @After
-    public void clearReplicationPeers() {
-        TestClusterStarter.clearReplicationPeers();
-        // TODO also cleanup data
-    }
+public class HBaseSourceITCase extends TestsWithTestHBaseCluster {
 
     @Test
     public void testBasicPut() throws Exception {
@@ -62,7 +34,7 @@ public class HBaseSourceITCase {
                         null,
                         deserializationScheme,
                         DemoSchema.TABLE_NAME,
-                        TestClusterStarter.getConfig());
+                        HBaseTestClusterUtil.getConfig());
         // NumberSequenceSource source = new NumberSequenceSource(1, 10);
         StreamExecutionEnvironment env = StreamExecutionEnvironment.getExecutionEnvironment();
         env.setParallelism(1);
@@ -75,23 +47,30 @@ public class HBaseSourceITCase {
         DemoIngester ingester = new DemoIngester();
         Tuple2<Put, String[]> put = ingester.createPut();
         String[] expectedValues = put.f1;
-        List<String> collectedValues = new ArrayList<>();
+        expectFirstValuesToBe(
+                stream,
+                expectedValues,
+                "HBase source did not produce the right values after a basic put operation");
+
+        doAndWaitForSuccess(env, () -> ingester.commitPut(put.f0), 120);
+    }
+
+    private static <T> void expectFirstValuesToBe(
+            DataStream<T> stream, T[] expectedValues, String message) {
+
+        List<T> collectedValues = new ArrayList<>();
         stream.flatMap(
-                new RichFlatMapFunction<String, Object>() {
+                new RichFlatMapFunction<T, Object>() {
 
                     @Override
-                    public void flatMap(String value, Collector<Object> out) throws Exception {
+                    public void flatMap(T value, Collector<Object> out) {
                         collectedValues.add(value);
                         if (collectedValues.size() == expectedValues.length) {
-                            assertArrayEquals(
-                                    "HBase source did not produce the right values after a basic put operation",
-                                    expectedValues,
-                                    collectedValues.toArray());
+                            assertArrayEquals(message, expectedValues, collectedValues.toArray());
                             throw new SuccessException();
                         }
                     }
                 });
-        doAndWaitForSuccess(env, () -> ingester.commitPut(put.f0), 60);
     }
 
     private static void doAndWaitForSuccess(
@@ -109,14 +88,6 @@ public class HBaseSourceITCase {
                 // Normal termination
             }
         }
-    }
-
-    private static boolean causedBySuccess(Exception exception) {
-        boolean success = false;
-        for (Throwable e = exception; !success && e != null; e = e.getCause()) {
-            success = success || e instanceof SuccessException;
-        }
-        return success;
     }
 
     /** Bla. */
